fix(doctor): parse v2 JSON lockfile instead of flagging it unparseable
tandem doctor's annotation store-lock check still read the lock as a bare
PID via parseInt, but locks have been v2 JSON ({pid, startedAtMs, app})
since #1077. A healthy lock therefore produced a spurious warning
('unparseable content: {...}'), seen during the v0.14.0 npm smoke pass.

Extract the pure lockfile format/parsing (parseLockfile, lockfilePayload,
LockfileContents) into src/server/annotations/lockfile.ts so the CLI doctor
can reuse it without importing the store's file-io/notifications/platform
graph; store.ts re-exports parseLockfile/LockfileContents unchanged (the
existing test import path is preserved). The doctor now delegates to
parseLockfile, which reads both v2 JSON and legacy raw-PID locks.

Verified: typecheck clean, store + diagnostics tests pass, and a real
v2-JSON lock now reports holder PID + liveness instead of 'unparseable'.
